+++ src/components/auth/join/JoinComponent.tsx
@@ -66,11 +66,11 @@ const JoinComponent = () => {
     setPassword("");
 
     try {
-      const { data, error } = await supabase.auth.signUp({
+      const { data, error: signUpError } = await supabase.auth.signUp({
         email: formdata.email,
         password: formdata.password,
       });
-
+      console.log(signUpError);
       const userId = async () => {
         const {
           data: { user },
@@ -85,7 +85,6 @@ const JoinComponent = () => {
     setFreeLancerEmail(false);
     setShowFreeLancerJoin(true);
   };
-
   const handleClientJoin = () => {
     setShowFreeLancerJoin(false);
     setShowClientEmail(!showClientEmail);
@@ -184,15 +183,15 @@ const JoinComponent = () => {
                 placeholder="비밀번호"
                 {...register("password", {
                   required: true,
-                  minLength: 6,
+                  // minLength: 6,
                 })}
               />
               {errors.password && errors.password.type === "required" && (
                 <p>비밀번호를 입력하세요</p>
               )}
-              {errors.password && errors.password.type === "minLength" && (
+              {/* {errors.password && errors.password.type === "minLength" && (
                 <p>비밀번호는 최소 6자리 이상</p>
-              )}
+              )} */}
               <button>다음</button>
             </div>
           </form>
@@ -224,15 +223,15 @@ const JoinComponent = () => {
                 placeholder="비밀번호"
                 {...register("password", {
                   required: true,
-                  minLength: 6,
+                  // minLength: 6,
                 })}
               />
               {errors.password && errors.password.type === "required" && (
                 <p>비밀번호를 입력하세요</p>
               )}
-              {errors.password && errors.password.type === "minLength" && (
+              {/* {errors.password && errors.password.type === "minLength" && (
                 <p>비밀번호는 최소 6자리 이상</p>
-              )}
+              )} */}
               <button>다음</button>
             </div>
           </form>
